Create a file for each decorator and import
