+++ decorator.rb
@@ -10,15 +10,3 @@ class Decorator < Nameable
     @nameable.correct_name
   end
 end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.upcase
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.length > 10 ? @nameable.correct_name.slice(0, 10) : @nameable.correct_name
-  end
-end
