+++ person.rb
@@ -1,5 +1,7 @@
 require_relative 'nameable'
 require_relative 'decorator'
+require_relative 'capitalizedecorator'
+require_relative 'trimmerdecorator'
 
 class Person < Nameable
   attr_accessor :name, :age
@@ -34,5 +36,5 @@ person = Person.new(22, 'maximilianus')
 person.correct_name
 capitalized_person = CapitalizeDecorator.new(person)
 capitalized_person.correct_name
-capitalized_trimmed_person = TrimmerDecorator.new(capitalizedPerson)
+capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
 puts capitalized_trimmed_person.correct_name
